feat: allow classes and strings for torch params
All torch-related params of estimators can now be torch classes or strings.
