+++ river_torch/anomaly/rolling_ae.py
@@ -26,7 +26,7 @@ class RollingWindowAutoencoder(base.Autoencoder):
         encoder_fn,
         decoder_fn,
         loss_fn="smooth_mae",
-        optimizer_fn: Type[torch.optim.Optimizer] = "sgd",
+        optimizer_fn = "sgd",
         device="cpu",
         window_size=50,
         **net_params,
