+++ river_torch/anomaly/variational_ae.py
@@ -25,7 +25,7 @@ class VariationalAutoencoder(base.Autoencoder):
         encoder_fn,
         decoder_fn,
         loss_fn="smooth_mae",
-        optimizer_fn: Type[torch.optim.Optimizer] = "sgd",
+        optimizer_fn= "sgd",
         device="cpu",
         n_reconstructions=10,
         beta=1,
