+++ river_torch/classification/classifier.py
@@ -55,7 +55,7 @@ class Classifier(DeepEstimator, base.Classifier):
         self,
         build_fn,
         loss_fn: str = "ce",
-        optimizer_fn: typing.Type[torch.optim.Optimizer] = torch.optim.SGD,
+        optimizer_fn="sgd",
         learning_rate=1e-3,
         device="cpu",
         seed=42,
@@ -175,7 +175,7 @@ class RollingClassifier(RollingDeepEstimator, base.Classifier):
         self,
         build_fn,
         loss_fn: str = "ce",
-        optimizer_fn: typing.Type[torch.optim.Optimizer] = torch.optim.SGD,
+        optimizer_fn= "sgd",
         window_size=1,
         device="cpu",
         learning_rate=1e-3,
