+++ river_torch/utils/__init__.py
@@ -1,7 +1,7 @@
 """Utility classes and functions."""
 from .estimator_checks import check_estimator
 from .layers import SequentialLSTM
-from .module_finder import get_activation_fn, get_init_fn, get_loss_fn, get_optimizer_fn
+from .params import get_activation_fn, get_init_fn, get_loss_fn, get_optim_fn
 from .river_compat import dict2tensor
 
 __all__ = [
@@ -9,7 +9,7 @@ __all__ = [
     "SequentialLSTM",
     "get_activation_fn",
     "get_loss_fn",
-    "get_optimizer_fn",
+    "get_optim_fn",
     "get_init_fn",
     "dict2tensor",
 ]
